pyproject.toml
[tool.poetry]
name = "os-computer-use"
version = "0.1.0"
description = ""
authors = ["James Murdza <james@jamesmurdza.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
e2b = "^1.5.0"
ipython = "^8.29.0"
python-dotenv = "^1.0.1"
gradio-client = "^1.5.0"
pillow = "^11.0.0"
e2b-desktop = "^1.7.2"
openai = "^1.59.5"
anthropic = "^0.44.0"
pyqtwebengine = "^5.15.7"
pywebview = {extras = ["qt"], version = "^5.4"}
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-multipart = "^0.0.12"

[tool.poetry.scripts]
start = "main:main"
api = "api_server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
